+++ autosubmit/database/db_common.py
@@ -705,7 +705,7 @@ def _get_sqlalchemy_conn() -> Connection:
     can use a context-manager and keep the previous behaviour
     intact.
     """
-    return session.create_engine(BasicConfig.DATABASE_CONN_URL).connect()
+    return session.get_engine(db_path=Path(BasicConfig.DB_PATH)).connect()
 
 
 def _create_db_pg() -> bool:
@@ -912,14 +912,3 @@ def _get_experiment_id_sqlalchemy(name: str) -> int:
         raise AutosubmitCritical(f'The experiment "{name}" does not exist', 7005)
 
     return int(row[0])
-
-
-def get_connection_url(db_path: Optional['Path'] = None) -> str:
-    """Return a SQLAlchemy connection URL."""
-    if BasicConfig.DATABASE_BACKEND == "postgres":
-        return BasicConfig.DATABASE_CONN_URL
-
-    if not db_path:
-        raise ValueError('For SQLite databases you MUST provide a database file.')
-
-    return f'sqlite:///{str(Path(db_path).resolve())}'